feat: finalize admin CRUD navigation, real-time support, and recycle bin

+++ backend/src/main/java/com/example/DATN/dto/UserVideoChannelDto.java
@@ -9,7 +9,7 @@ public record UserVideoChannelDto(
         String url,
         String handle,
         String description,
-        Integer subscriberCount,
+        Long subscriberCount,
         Long videoCount
 ) {
 }
