+++ backend/src/main/java/com/example/DATN/repository/SupportTicketRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface SupportTicketRepository extends JpaRepository<SupportTicket, Integer> {
+public interface SupportTicketRepository extends JpaRepository<SupportTicket, Long> {
     @Query("""
             select st
             from SupportTicket st
